+++ src/db_migrate.cpp
@@ -13,7 +13,7 @@
 namespace
 {
 
-constexpr auto version = std::string_view{"0.5.36"};
+constexpr auto version = std::string_view{"0.5.37"};
 
 auto print_help() -> void
 {